refactor(escrow,token): split modules, add incident-response doc and monitor script
#535 — Refactor escrow/src/lib.rs into focused sub-modules
- Extract lifecycle functions (initialize, fund, mark_delivered, approve_delivery,
  release_partial, request_refund, cancel, extend_deadline, release_to_seller,
  refund_to_buyer, update_amount) into contracts/escrow/src/lifecycle.rs
- Extract dispute functions (raise_dispute, resolve_dispute) into
  contracts/escrow/src/dispute.rs
- Extract query helpers (get_escrow_info, get_state, is_deadline_passed,
  get_remaining_ledgers, bump) into contracts/escrow/src/queries.rs
- lib.rs now acts as a thin dispatcher; all public API signatures are unchanged

#536 — Extract TokenInterface impl into token/src/token_interface.rs
- Move the implementation bodies for all token::TokenInterface methods
  (allowance, approve, balance, transfer, transfer_from, burn, burn_from,
  decimals, name, symbol) into contracts/token/src/token_interface.rs
- lib.rs retains the #[contractimpl] block and delegates each method to
  the corresponding function in the new module
- Custom methods (mint, batch_mint, admin_burn, propose_admin, accept_admin,
  cancel_admin_proposal, set_admin, admin, total_supply, balance_of, version,
  contract_version, allowance_expiry) remain in lib.rs

#534 — Add docs/incident-response.md
- Covers: how to pause the contract immediately, how to rotate the admin key
  (two-step propose/accept flow), how to upgrade to a patched WASM (with
  timelock steps), how to communicate with affected users, post-incident
  review checklist

#533 — Add scripts/monitor-token.sh and update docs/deployment-guide.md
- scripts/monitor-token.sh [network]: prints admin, total supply, max supply,
  paused state, and contract version for a given TOKEN_CONTRACT_ID
- docs/deployment-guide.md: added §14 documenting the monitor script, its
  output format, and available fields

Closes #535
Closes #536
Closes #534
Closes #533
